Adds dev-mode script to allow testing locally
Adds dev-mode script which uses python-ircd to start a local irc
server and runs a bot on channel #test. Uses different colour output
for bot and ircd messages. Ideally I wanted these side by side like
two columns but this will have to do for now.
The pre-req to run the scripts is to have run the tests once, as it
installs python-ircd in the same "tests-env" virtualenv.