Changes buffer error handling from replace
Changes decoder error handling from strict to replace, therefore
replacing any bytes that failed to decode with "?".
This previously raised a:
  UnicodeDecodeError: 'utf8' codec can't decode byte 0xc2 in
  position 509: unexpected end of data
and complained about message size being bigger than 512.
This may not be the best fix for this, it might be better to just
strip out the bytes that we can't decode, as usually it will be
the last bytes where a multi-byte unicode has been chipped, but
doing that could lead to scenarios where messages are getting
clipped and might be much harder to debug.