Exception handling around reddit commands
Fixes:
Looks like while porting to python3 a lot of exceptions were mistakenly changed
and ValueErrors stopped being handled. Diff example:

https://github.com/apalos/tatianna/commit/f2a406e904d8ba8cf6c3baa66f91f789cc3ac994

I 've amended the ValueErrors when necessary. Added a check to fix the randrange
exception as well when there's only a single node in json returned. Added a
generic exception handler to the reddit command, since if it fails it is not
crucial and should allow the bot to continue as usual. Finally, added a class
initialiser for reddit feed that only gets initialised once, so that we
don't parse the subreddits file every time.